rename python package
