+++ cell_feature_data/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 """
-The `python_dataset_creator` package provides functionality for creating datasets in Python.
+The `cell_feature_data` package provides functionality for creating datasets in Python.
 
 Note:
 This package is still under development and may not have all the desired features. Please refer to the documentation for more information on each module and function.
